refactor(plan_SERVERSPLIT slice 3): the debugging bodies leave server.py
Five more bodies move into embedded/tool_sections/debugging.py, and this is the
first slice that RELOCATES rather than only moves.
validate_browser_environment_tool was filed three hundred lines away among the
element-extraction bodies while registering into `debugging` all along; it joins
its own section here. That is what makes module-to-section a clean 1:1, which is
what lets SECTION be a per-module constant instead of a per-tool string. A pure
relocation: no rename, no reordering of SECTION_TOOLS["debugging"], no behaviour
change, and the served surface is unmoved.

It is also the first slice to move a string that a source-TEXT guard is pinned to:
MSG_EXPORT_TIMEOUT lives in export_debug_logs, and "Export timeout - file too
large" now appears in tool_sections/debugging.py and nowhere in server.py. Slice
0's widening of that pin to source_scan.tool_source_text() is what carries the
guard into the new file. An inspect.getsource(server) pin would not have failed
here; it would have passed, over a file the message had already left. Verified
directly rather than assumed.

Verified: 2334 passed / 1 skipped on the unit lane; tool surface IDENTICAL to the
HARD golden; SECTION_TOOLS 94 with per-app [94, 94, 94]; --list-sections
subprocess "Total: 94 tools"; the derived tool-source set is now the five files it
should be; gen_release_contract --check, ruff format/check, vulture, ty and the
suppression-owner gate all clean; test_e2e_transport.py green on real Chrome over
real stdio.

server.py: 3144 to 3014 LOC, cap ratcheted DOWN to the measured actual, and the
two platform_utils imports the stray was the last consumer of are pruned with it.
tests/source_scan.py's floor ratchets UP 4 to 5. The plan's execution-log table
gains the seven drifts these three slices hit.
