+++ tools/check_file_budgets.py
@@ -66,8 +66,16 @@ GRANDFATHER: dict[str, tuple[int, str]] = {
     # (103 lines) moved VERBATIM into embedded/tool_sections/tabs.py, and the
     # one-line `_require_browser` import they were the last consumer of was
     # pruned with them. Cap == the measured post-ruff-format actual, no padding.
+    # plan_SERVERSPLIT slice 3 RATCHETS DOWN 3144 -> 3014. The five debugging
+    # bodies (128 lines) moved VERBATIM into embedded/tool_sections/debugging.py
+    # — including validate_browser_environment_tool, which was physically filed
+    # three hundred lines away among the element-extraction bodies while
+    # registering into `debugging`; it is RELOCATED to its own section, which is
+    # what makes module <-> section a clean 1:1. The two platform_utils imports
+    # it was the last consumer of were pruned with it. Cap == the measured
+    # post-ruff-format actual, no padding.
     "embedded/server.py": (
-        3144,
+        3014,
         "plan_M4ph1 + plan_M3 + plan_M10a + plan_F808 + plan_F809 + plan_SERVERSPLIT",
     ),
     # plan_M4ph1 C1 (F-201): the verbatim 50-def clone-storage move is an
